Refactor and cleanup
This commit reorders the appearance of function implementations, so it
looks a bit messy and larger than it is.

The most significant and real change is refactoring
`lxt_resolve_container` into a separate function `lxt_append_entry`
which is much more indicative of what actually happens. “Resolving a
container” does not make sense.