Add function to validate token identifiers
This adds validation when parsing container, generator and sequence
variable tokens, ensuring that container and generator names follow the
same constraints as imposed on sequence variables, mitigating the
potential for unresolvable references.

For example, until now, a container could be named whatever you like.
It could even just be an emoji. However, a container named e.g. "my
container" (note the whitespace) could never be resolved when
referenced as a sequence variable ("@my container") due to the nature
of how sequences are parsed. It would parse "my" and think that was a
reference to a container named "my".

I would prefer not putting in limits like this, but it seems like a
sensible thing to do to solve this issue.

See #2